Fixed incomplete <type_trait> test
Travis logged failed compilation test for the correctly implemented libc++ and
libstdc++. Test was missing the -std=c++14 flag. Still, a std::enable_if
solution might be favourable and could simply assume the worst case of
is_trivially_* = false.

In addition, Boost::Hana is tested on MSYS2/MinGW64 with
- Clang 3.6.2 / libstdc++ of GCC 5.2.0 / with exceptions
  -> failure while linking test.foldable.fold_left_mcd.part6 (unknown reason)
- GCC 5.2.0 / libstdc++ of GCC 5.2.0 / with exceptions
  -> failure, as expected

Want me to add some compilation notices to the documention ? Now it is however
fairly simple to get it compiling, even on Windows.