[Travis] Test against Boost 1.65.1 and drop versions older than 1.63
This reduces the load on Travis. Since we don't advertise anywhere that Hana
works with older versions of Boost, if this ends up breaking someone (because
they're mixing Hana and Boost at different versions and we introduce a change
that breaks under some older Boost), it can be argued that it's their problem.