[config] Remove Standard Library detection
In most cases, this shouldn't be necessary if we detect the compiler
being used. Strictly speaking, this isn't true because someone could
override the Standard Library with any compiler. However, the logic
for detecting the Standard Library is broken for recent libc++'s, and
I don't want to maintain that.

Also, the Standard Libraries that are not supported are getting more
and more ancient, so the benefit of having these warnings in place is
getting smaller and smaller.