The base branch is now master, and all releases are done through the Boost release process
Given the lack of bandwidth and the infrequency at which I get to look
at issues in this repo, simplifying the process of fixing stuff and
releasing is actually important. From now on, we'll just merge stuff
to master and it'll always be ready to be picked up by the next Boost
release.