Re-add MSVC workarounds that were dropped by e315075
e315075 re-generated the `struct_macros.hpp` file, which had the
unfortunate side effect of dropping MSVC workarounds that had
been added in 95f8f74 because those were not added to the `.erb` file.
This adds the workarounds to the `.erb` file and re-generates the `.hpp`
file to restore the workarounds.

Thanks to @Grumpfy for pointing this out in [1].

[1]: https://github.com/boostorg/hana/commit/e315075784745beadac0a73ba22da332b40801ae#commitcomment-104250393